Added EditorButtons component
